requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-whois==0.8.0
dnspython==2.4.2
requests==2.31.0
python-wappalyzer==0.3.1
fpdf==1.7.2
cryptography==41.0.5
tldextract==5.1.1
setuptools